requirements.txt
pandas
numpy
wordcloud
cv2
os
requests
scipy
fpdf
Pillow
genanki
matplotlib
barcode
pymongo
fastapi
pydantic
uvicorn
bson
pydantic[email]
pyjwt
fsrs
bcrypt
motor